requirements.txt
requests>=2.28.0
pyyaml>=6.0
openai>=1.0.0
ollama>=0.1.7
argparse>=1.4.0
pathlib>=1.0.1
subprocess32>=3.5.4
jinja2>=3.1.0
colorama>=0.4.6 